Move application class to separate packege (#288)
