+++ embedded-oracle-xe/src/test/java/com/playtika/test/oracle/dummy/TestApplication.java
@@ -1,4 +1,4 @@
-package com.playtika.test.oracle;
+package com.playtika.test.oracle.dummy;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
